fix: hide netrw banner by default, so user can jump using path number prefixes
Author uses sequential number prefixes so I can easily jump to and open
files. E.g., I have a folder of symlinks to notes files that I open
frequently, and they're numbered from "01" .. "35". When I open netrw, the
number prefix matches the number of lines from the top of the netrw buffer.
So, for instance, if a file's prefix is "05", I can press 5<Down><CR>
to quickly open that file (in the same window as the netrw buffer,
which then closes).

- This has been working for me because my Neovim config disables the
  netrw banner from another plugin, but I noticed it was missing from
  this plugin.